Complete exercise 3.12
Created a new .js file to practice setting up a MongoDB database.
The functionality of this script is to allow for viewing all the
phonebook entries in the database if the command line args only include
the password. However, if the command line args also include a name
and a number, the program will instantiate a new person object to be
saved to the database.