Complete exercise 3.17 and 3.18
Added functionality that allows the user to update the database
with a new number if the input name already exists in the database.
When the user is updating a name, the backend will make a put request
to the database to update the number for the specified person id.
Also updated the route for /info to use the database to determine
the information.